add_neighbors now sorts, finds dist and angles
For each particle i,Makes a list of tuples, with:
    (id_j, distance_ij, angle_ij)
for each neighbor j.
Sorts by distance, and keeps closest nn=6 particles.

Need some method to store these tuples, should still be able to use the
structured array, but now it must be a list of tuples with their own
dtype instead of a list of ints.

	modified:   correlation.py