Lots of changes, mainly: killjump, singletracks
To implement singletracks for the MSD plots, new information must be
saved in the msd.npz files, in a new array called msdids which contains
the track ids for each track msd, which was otherwise lost.
singletracks now works for all of plotmsd, plottracks, plotorient.

kill_jumps and kill_flats were modified to just look at the beginning
and end (not so much averaging, no longer a factor for kill_flats)

some more print information (names files, e.g.)
better verbosity control

as of this commit, the msd magnitude is not known to be correct.

	modified:   otracks.py
	modified:   tracks.py