New option --cut to cut tracks at boundary
When a particle comes near the boundary (center within one particle
width) then its track will be cut and when it leaves the boundary it is
given a new trackid. Particles near the boundary, as stubs before it,
are given the trackid = -1, which are later ignored. If a later particle
has a parent with trackid -1, it gets a new trackid.

Boundary location is determined by three clicks along the boundary which
defines a circle. User must do this interactively.

	modified:   helpy.py
	modified:   tracks.py