New function to remove_duplicates from tracks
Sometimes, two dots in a single frame get assigned to the same track.
This then gives a non-increasing time for the track (since it's in two
places at once). After tracks have been found, go through them all and
check for this problem, keeping only the track that's closest, in sum,
to the positions in the previous and following frames.

This function can also be used as a standalone function on old tracked
data. No need to rerun the full tracking process.