Plot histograms when running positions
Mostly useful when trying to figure out the paramters for particle
detection. Saves a histogram of the sizes and eccentricities of each
segment, both corner and center if used.

Also save the segment plots to SEGMENTS instead of POSITIONS, and use
marker sizes that scale with the kernel size in the scatter plots.

Rearranged a little code related to the plotting, but nothing
substantial there.