Progress on refactoring fitting and plotting corrs
Cannot have a single function for all curves, too much stuff is too
specific to the correlation function, and with lmfit the actual fitting
and parameter management is quite simple, so all the actual code is
specifying it for the particular correlation.

Looping through and printing the parameter fits, as well as plotting
them, is the real duplicated stuff. Here I've added a fit_format
function that formats the results to print and save in various places.