Revert "Make pair_indices same shape if asarray or not"
This reverts commit 75375d08b28236925a474e2938179dfa866d085e.
I am not sure what the original purpose of this commit was, other than
some sort of symmetry. But I believe that the original intent for the
asarray=True argument was to use the entire thing as an index, and this
way the first index gives the id of the pair, as opposed to the second.
The commit is being reverted because it caused problems in the
orientation.find_corners function.