BIG BUG FIX in units for Diffusion from Variance!
Was missing a unit conversion from variance to diffusion. Not a
mishandling of units, but rather a mishandling of assumption that in
variance = 2 * D * delta(t)
the delta(t) does not go to 1, because we've actually *added* units, not
*removed* them, by going to natural physical dimensions instead of array
stepsize dimensions, so now delta(t) = 1/dt = fps, instead of 1.

Note that this is something that would have been impossible to fix in
the handling of units at the level of velocity derivatives or position
measurements. It must happen after the variance is taken.