Subtract cumulant as "mean" or "initial" properly
Cumulant = [bool, bool] just subtracted the total mean of the arrays
before correlating, but the mean is a function of shift dx, as it should
only be averaged over the valid (overlapping) parts of the array. This
simultaneously corrects subtracting out the initial point (dx = 0),
which is also different at different values of dx. Instead of
cumulant=True use cumulant='mean', instead of norm=0 use cumulant='init'